requirements.txt
beautifulsoup4==4.12.3
easyocr==1.7.2
imbalanced_learn==0.12.4
imblearn==0.0
matplotlib==3.9.2
numpy==2.1.3
openpyxl==3.1.5
pandas==1.5.3
pdfplumber==0.9.0
Pillow==11.0.0
PyMuPDF==1.21.1
PyPDF2==3.0.1
python_docx==1.1.2
pywin32==308
scikit_learn==1.1.2
spacy==3.8.0
sparkai==0.3.0
tensorboard==2.14.1
tensorflow==2.14.0
tensorflow_intel==2.14.0
torch==2.0.1
tqdm==4.65.0
transformers==4.22.1
beautifulsoup4==4.12.3
matplotlib==3.9.2
numpy==1.21.0
Pillow==11.0.0
PyPDF2==3.0.1
